+++ lib/shrine/plugins/activerecord.rb
@@ -28,8 +28,10 @@ class Shrine
 
           if shrine_class.opts[:activerecord][:validations]
             model.validate do
-              send(:"#{name}_attacher").errors.each do |message|
-                errors.add(name, *message)
+              if send(:"#{name}_attacher").respond_to?(:errors)
+                send(:"#{name}_attacher").errors.each do |message|
+                  errors.add(name, *message)
+                end
               end
             end
           end