+++ test/plugin/activerecord_test.rb
@@ -26,6 +26,10 @@ describe Shrine::Plugins::Activerecord do
 
   describe "Attachment" do
     describe "validate" do
+      before do
+        @shrine.plugin :validation
+      end
+
       it "adds attacher errors to the record" do
         @user.class.include @shrine::Attachment.new(:avatar)